Fix the quantity form code issue in the bag page
In the shopping bag page, there are two occurences of quantity form,
for each product. One is visible and other is hidden, depending on
the screen size. This commit updates the quantity form id, making it
unique as required by the HTML. Then the JavaScript is updated to
handle both quantity forms that exist for each product, making sure
the plus or the minus buttons are disabled in the corresponding
quantity form, when the user tries to set a number less than 1 or
greater than 99.